change --everything opt, rework tests
