+++ .rubocop.yml
@@ -3,3 +3,6 @@ AllCops:
     - 'spec/**'
 Metrics/MethodLength:
   Max: 12
+Metrics/BlockLength:
+  Exclude:
+    - 'bin/**'
